Commit changes for V1.85.3
- Added multiple extra API checks when a game's ".bin" file is updated in order to minimise the risk that an achievement unlock will be missed if the API is not updated before the check occurs.
- Updated logos to SVG format.
- Prevented further achievement tracking for games with 100% completion, and added a gold colour to the game's name in the "Game Display" box to indicate all achievements have been earned.
- Fixed an issue where the screenshot background window could take focus from the game when the "Screenshot with Notification Overlay" option was enabled.
- Fixed an issue where the application would stop tracking stats after launching a game with no achievements.
- Fixed an issue where continuous notifications would spawn after launching a game that was previously completed.
- Fixed an issue where the "Use Game Icon" Customiser option did not function correctly.
- Fixed an issue where percentages for the "Epic Games" preset did not show the actual value.
- Fixed an issue where the "Game Display" box would sometimes retain the previous game name instead of resetting back to the "No Game Detected" text after closing a game.
- Set the "window-state" plugin to only read the "SIZE" and "POSITION" properties of the main window on launch, which prevents an issue where a previously maximised window would stay stuck in a maximised state upon each subsequent launch of the application.

+++ src-tauri/src/main.js
@@ -83,17 +83,19 @@ window.addEventListener("DOMContentLoaded", async () => {
         log.write("info",event.payload.msg)
 
         if (parseInt(event.payload.optional) === 0) {
-            
             async function CheckQueueIsEmpty() {
                 if (!queue.length) {
                     window.gameName = null
-                    
-                    gamelbl.textContent = translations.gamelbl.elem
-                    gamelbl.setAttribute("novalue","")
 
-                    await removeDir(await path.join(cachepath,appid), { recursive: true })
+                    appid && await removeDir(await path.join(cachepath,appid), { recursive: true })
                     window.appid = null
                     delete window.cachedata
+                    
+                    gamelbl.setAttribute("novalue","")
+                    gamelbl.removeAttribute("complete")
+
+                    // Resets "gamelbl" variable back to default value if gameName is null
+                    LoadLang(config.lang)
                 } else {
                     console.log("%cWaiting for queue to be empty...","color:darkred")
                     return setTimeout(CheckQueueIsEmpty,1000)
@@ -133,11 +135,12 @@ window.addEventListener("DOMContentLoaded", async () => {
         })
         .finally(async () => {
             try {
-                window.appid = event.payload.optional
-                
                 const { gameName, cachedata, gameicon } = await CacheAchievementData(event.payload.optional)
+
+                window.appid = event.payload.optional
                 window.gameName = gameName
                 window.cachedata = cachedata
+                window.gameicon = gameicon
     
                 config.track && ShowInfoNotify(gameName,gameicon ? convertFileSrc(gameicon) : "../img/sanlogo.svg")
     
@@ -146,8 +149,20 @@ window.addEventListener("DOMContentLoaded", async () => {
                 gamelbl.textContent = gameName
                 gamelbl.removeAttribute("novalue")
             } catch (err) {
-                log.write("error",`Error caching achievement data: ${err}`)
-                ShowInfoNotify(translations.trackingfailed,"../icon/error.svg",translations.checkapplog)
+                if (err === 0) {
+                    log.write("info",`AppID ${event.payload.optional} has no achievements`)
+                } else if (err[0] === 100) {
+                    log.write("info",`AppID ${event.payload.optional} has all achievements unlocked`)
+
+                    window.gameName = err[1]
+
+                    gamelbl.textContent = gameName
+                    gamelbl.removeAttribute("novalue")
+                    gamelbl.setAttribute("complete","")
+                } else {
+                    log.write("error",`Error caching achievement data: ${err}`)
+                    ShowInfoNotify(translations.trackingfailed,"../icon/error.svg",translations.checkapplog)
+                }
             }
         })
     })
@@ -191,8 +206,14 @@ window.addEventListener("DOMContentLoaded", async () => {
     })
 
     listen("modified",event => {
-        cachedata ? CheckUnlockStatus(event.payload.optional) : log.write("cache miss",`"cachedata" was missing on last check`)
-        log.write("info",`MSG: ${event.payload.msg}\nAPPID: ${event.payload.optional}`)
+        const checkunlockstatus = msg => {
+            cachedata ? CheckUnlockStatus(event.payload.optional) : log.write("cache miss",`"cachedata" was missing on last check`)
+            log.write("info",msg)
+        }
+
+        checkunlockstatus(`MSG: ${event.payload.msg}\nAPPID: ${event.payload.optional}`)
+        setTimeout(() => checkunlockstatus(`Re-checking unlock status for APPID: ${event.payload.optional}`),1000)
+        setTimeout(() => checkunlockstatus(`Final re-check of unlock status for APPID: ${event.payload.optional}`),2000)
     })
 
     listen("notifyerror", event => {
